Long and fast pipes

+++ tests/System.IO.Pipelines.Tests/WritableBufferFacts.cs
@@ -94,14 +94,14 @@ namespace System.IO.Pipelines.Tests
                 await output.FlushAsync();
                 pipe.Writer.Complete();
 
-                int offset = 0;
+                long offset = 0;
                 while (true)
                 {
                     var result = await pipe.Reader.ReadAsync();
                     var input = result.Buffer;
                     if (input.Length == 0) break;
-
-                    Assert.True(input.EqualsTo(new Span<byte>(data, offset, input.Length)));
+                    // Sizes less then max int
+                    Assert.True(input.EqualsTo(new Span<byte>(data, (int) offset, (int) input.Length)));
                     offset += input.Length;
                     pipe.Advance(input.End);
                 }
@@ -129,7 +129,7 @@ namespace System.IO.Pipelines.Tests
                 await output.FlushAsync();
                 pipe.Writer.Complete();
 
-                int offset = 0;
+                long offset = 0;
                 while (true)
                 {
                     var result = await pipe.Reader.ReadAsync();
@@ -137,7 +137,8 @@ namespace System.IO.Pipelines.Tests
                     if (input.Length == 0) break;
 
                     string s = ReadableBufferExtensions.GetUtf8String(input);
-                    Assert.Equal(data.Substring(offset, input.Length), s);
+                    // Sizes less then max int
+                    Assert.Equal(data.Substring((int) offset, (int) input.Length), s);
                     offset += input.Length;
                     pipe.Advance(input.End);
                 }
@@ -164,7 +165,7 @@ namespace System.IO.Pipelines.Tests
                 await output.FlushAsync();
                 pipe.Writer.Complete();
 
-                int offset = 0;
+                long offset = 0;
                 while (true)
                 {
                     var result = await pipe.Reader.ReadAsync();
@@ -172,7 +173,7 @@ namespace System.IO.Pipelines.Tests
                     if (input.Length == 0) break;
 
                     string s = ReadableBufferExtensions.GetAsciiString(input);
-                    Assert.Equal(data.Substring(offset, input.Length), s);
+                    Assert.Equal(data.Substring((int) offset, (int) input.Length), s);
                     offset += input.Length;
                     pipe.Advance(input.End);
                 }
